:racehorse: speed up pages by compressing HTML
Uses [a nifty Jekyll layout](http://jch.penibelst.de/) to compress
html. This cuts this bytes off my HTML document by 1/3 or more.

If we take GZIP into account (and we should, because github pages uses
it), the savings is probably a lot less (see
https://css-tricks.com/the-difference-between-minification-and-gzipping/#article-header-id-2)
but it's still worth doing because there will be tangible savings.

The compressed html is less readable, but it gets reconstructed in the
Devtools Elements pane, and it's still technically valid.