Set up logo testing page
See #25

Changes the logo setup to be a single inline SVG with a random script file instead, including a test page.

(I gave up trying to build `jekyll-gh-pages` locally after nokogiri complained about libxml2 for the second hour in succession. This is an attempt to make my repo on GitHub build it, since this is all their fault anyway.)