fix: use 'last_modified_at' instead of 'updated' in frontmatter
The previous variable 'updated' was arbitrarily named, but I recently
learned that the jekyll_sitemap plugin looks for a variable named
'last_modified_at', when building the sitemap. It was straightforward
for me to switch everything over to that name, and now my sitemap
reports updates to my posts.

This commit also includes an update to my now page. I probably should
have put that in a separate commit but oh well.